Fix infinite watcher recursion
Make a copy of the layout before assigning it to the graph
element. Plotly sometimes performs its own updates on the layout
element, specifically when dealing with time series data where
the x-axis consists of Date values. In this case, we can get
into an infinite recursion situation with the watcher in
angular.